Fix quality score calculations
When processing a pileup file, if a site has no data, then all subsequent sites will receive a quality score of -2.

In the refCalc function, if a site has no data the flag calc_rq_flag is set to False. However, this flag is never reset, so all following sites will have a quality score of -2.